@@ -1,15 +1,14 @@
 name: flutter_map_dragmarker
-description: Dragmarker class for flutter_map
+description: Dragmarker class for flutter_map null safety
 
-version: 1.0.0+1
+version: 1.0.0+2
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  cupertino_icons: any
   flutter_map: ^0.13.1
   latlong2: any
 
