@@ -6,26 +6,25 @@ import 'package:flutter_map/plugin_api.dart';
 
 class DragMarkerPluginOptions extends LayerOptions {
   List<DragMarker> markers;
-  DragMarkerPluginOptions({this.markers});
+  DragMarkerPluginOptions({this.markers = const []});
 }
 
 class DragMarkerPlugin implements MapPlugin {
   @override
-  Widget createLayer(
-      LayerOptions options, MapState mapState, Stream<Null> stream) {
+  Widget createLayer(options, mapState, stream) {
     if (options is DragMarkerPluginOptions) {
-      return StreamBuilder<int>(
+      return StreamBuilder<Null>(
           stream: stream,
-          builder: (BuildContext context, AsyncSnapshot<int> snapshot) {
+          builder: (context, snapshot) {
             var dragMarkers = <Widget>[];
             for (var marker in options.markers) {
               if (!_boundsContainsMarker(mapState, marker)) continue;
-
               dragMarkers.add(DragMarkerWidget(
-                  mapState: mapState,
-                  marker: marker,
-                  stream: stream,
-                  options: options));
+                mapState: mapState,
+                marker: marker,
+                stream: stream,
+                options: options,
+              ));
             }
             return Container(
               child: Stack(children: dragMarkers),
@@ -56,36 +55,32 @@ class DragMarkerPlugin implements MapPlugin {
 }
 
 class DragMarkerWidget extends StatefulWidget {
-  DragMarkerWidget(
-      {this.mapState,
-      this.marker,
-      AnchorPos anchorPos,
-      this.stream,
-      this.options}); //: anchor = Anchor.forPos(anchorPos, marker.width, marker.height);
+  DragMarkerWidget({
+    required this.mapState,
+    required this.marker,
+    AnchorPos? anchorPos,
+    required this.stream,
+    required this.options,
+  }) : anchor = Anchor.forPos(anchorPos, marker.width, marker.height);
 
   final MapState mapState;
-  //final Anchor anchor;
   final DragMarker marker;
   final Stream<Null> stream;
   final LayerOptions options;
+  final Anchor? anchor;
 
   @override
   _DragMarkerWidgetState createState() => _DragMarkerWidgetState();
 }
 
 class _DragMarkerWidgetState extends State<DragMarkerWidget> {
-  CustomPoint pixelPosition;
-  LatLng dragPosStart;
-  LatLng markerPointStart;
-  LatLng oldDragPosition;
+  late CustomPoint pixelPosition;
+  late LatLng dragPosStart;
+  late LatLng markerPointStart;
+  late LatLng oldDragPosition;
   bool isDragging = false;
 
-  static Timer autoDragTimer;
-
-  @override
-  void initState() {
-    super.initState();
-  }
+  static Timer? autoDragTimer;
 
   @override
   Widget build(BuildContext context) {
@@ -97,10 +92,10 @@ class _DragMarkerWidgetState extends State<DragMarkerWidget> {
       onPanUpdate: onPanUpdate,
       onPanEnd: onPanEnd,
       onTap: () {
-        if (marker.onTap != null) marker.onTap(marker.point);
+        if (marker.onTap != null) marker.onTap!(marker.point);
       },
       onLongPress: () {
-        if (marker.onLongPress != null) marker.onLongPress(marker.point);
+        if (marker.onLongPress != null) marker.onLongPress!(marker.point);
       },
       child: Stack(children: [
         Positioned(
@@ -108,14 +103,14 @@ class _DragMarkerWidgetState extends State<DragMarkerWidget> {
           height: marker.height,
           left: pixelPosition.x +
               ((isDragging && (marker.feedbackOffset != null))
-                  ? marker.feedbackOffset.dx
+                  ? marker.feedbackOffset!.dx
                   : marker.offset.dx),
           top: pixelPosition.y +
               ((isDragging && (marker.feedbackOffset != null))
-                  ? marker.feedbackOffset.dy
+                  ? marker.feedbackOffset!.dy
                   : marker.offset.dy),
           child: (isDragging && (marker.feedbackBuilder != null))
-              ? marker.feedbackBuilder(context)
+              ? marker.feedbackBuilder!(context)
               : marker.builder(context),
         ),
       ]),
@@ -135,14 +130,14 @@ class _DragMarkerWidgetState extends State<DragMarkerWidget> {
         (pos.y - (marker.height - widget.marker.anchor.top)).toDouble());
   }
 
-  void onPanStart(details) {
+  void onPanStart(DragStartDetails details) {
     isDragging = true;
     dragPosStart = _offsetToCrs(details.localPosition);
     markerPointStart =
         LatLng(widget.marker.point.latitude, widget.marker.point.longitude);
 
     if (widget.marker.onDragStart != null)
-      widget.marker.onDragStart(details, widget.marker.point);
+      widget.marker.onDragStart!(details, widget.marker.point);
   }
 
   void onPanUpdate(DragUpdateDetails details) {
@@ -180,17 +175,17 @@ class _DragMarkerWidgetState extends State<DragMarkerWidget> {
       /// around a bit to keep it going.
 
       var lastTick = 0;
-      if (autoDragTimer != null) lastTick = autoDragTimer.tick;
+      if (autoDragTimer != null) lastTick = autoDragTimer!.tick;
 
       if ((autoOffsetY != 0.0) || (autoOffsetX != 0.0)) {
         adjustMapToMarker(widget, autoOffsetX, autoOffsetY);
 
-        if ((autoDragTimer == null || autoDragTimer.isActive == false) &&
+        if ((autoDragTimer == null || autoDragTimer!.isActive == false) &&
             (isDragging == true)) {
           autoDragTimer =
               Timer.periodic(const Duration(milliseconds: 10), (Timer t) {
-            if (isDragging == false || (autoDragTimer.tick > lastTick + 15)) {
-              autoDragTimer.cancel();
+            if (isDragging == false || (autoDragTimer!.tick > lastTick + 15)) {
+              autoDragTimer!.cancel();
             } else {
               /// Note, we may have adjusted a few lines up in same drag,
               /// so could test for whether we've just done that
@@ -209,7 +204,8 @@ class _DragMarkerWidgetState extends State<DragMarkerWidget> {
       updatePixelPos(marker.point);
     });
 
-    if (marker.onDragUpdate != null) marker.onDragUpdate(details, marker.point);
+    if (marker.onDragUpdate != null)
+      marker.onDragUpdate!(details, marker.point);
   }
 
   /// If dragging near edge of the screen, adjust the map so we keep dragging
@@ -225,14 +221,14 @@ class _DragMarkerWidgetState extends State<DragMarkerWidget> {
 
     marker.point = mapState.unproject(CustomPoint(
         oldMarkerPoint.x + autoOffsetX, oldMarkerPoint.y + autoOffsetY));
-    mapState.move(newMapLatLng, mapState.zoom);
+    mapState.move(newMapLatLng, mapState.zoom, source: MapEventSource.custom);
   }
 
-  void onPanEnd(details) {
+  void onPanEnd(DragEndDetails details) {
     isDragging = false;
-    if (autoDragTimer != null) autoDragTimer.cancel();
+    if (autoDragTimer != null) autoDragTimer!.cancel();
     if (widget.marker.onDragEnd != null)
-      widget.marker.onDragEnd(details, widget.marker.point);
+      widget.marker.onDragEnd!(details, widget.marker.point);
     setState(() {}); // Needed if using a feedback widget
   }
 
@@ -259,24 +255,24 @@ class _DragMarkerWidgetState extends State<DragMarkerWidget> {
 class DragMarker {
   LatLng point;
   final WidgetBuilder builder;
-  final WidgetBuilder feedbackBuilder;
+  final WidgetBuilder? feedbackBuilder;
   final double width;
   final double height;
   final Offset offset;
-  final Offset feedbackOffset;
-  final Function(DragStartDetails, LatLng) onDragStart;
-  final Function(DragUpdateDetails, LatLng) onDragUpdate;
-  final Function(DragEndDetails, LatLng) onDragEnd;
-  final Function(LatLng) onTap;
-  final Function(LatLng) onLongPress;
+  final Offset? feedbackOffset;
+  final Function(DragStartDetails, LatLng)? onDragStart;
+  final Function(DragUpdateDetails, LatLng)? onDragUpdate;
+  final Function(DragEndDetails, LatLng)? onDragEnd;
+  final Function(LatLng)? onTap;
+  final Function(LatLng)? onLongPress;
   final bool updateMapNearEdge;
   final double nearEdgeRatio;
   final double nearEdgeSpeed;
   Anchor anchor;
 
   DragMarker({
-    this.point,
-    this.builder,
+    required this.point,
+    required this.builder,
     this.feedbackBuilder,
     this.width = 30.0,
     this.height = 30.0,
@@ -290,8 +286,6 @@ class DragMarker {
     this.updateMapNearEdge = false, // experimental
     this.nearEdgeRatio = 1.5,
     this.nearEdgeSpeed = 1.0,
-    AnchorPos anchorPos,
-  }) {
-    anchor = Anchor.forPos(anchorPos, width, height);
-  }
+    AnchorPos? anchorPos,
+  }) : this.anchor = Anchor.forPos(anchorPos, width, height);
 }
